requirements.txt
# apt-clone==0.2.1
# apturl==0.5.2
asgiref==3.7.2
beautifulsoup4==4.10.0
blinker==1.4
# Brlapi==0.8.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
cloudinary==1.34.0
# colorama==0.4.4
# command-not-found==0.3
configobj==5.0.6
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==41.0.1
# cupshelpers==1.0
dbus-python==1.2.18
# defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
dj-database-url==2.1.0
Django==4.2.3
django-admin-volt==1.0.10
django-allauth==0.54.0
django-ckeditor==6.6.1
django-crispy-forms==2.0
django-fontawesome-5==1.0.18
django-jazzmin==2.6.0
django-jet-4==0.0.1
django-js-asset==2.1.0
django-ninja==0.22.2
django-simple-history==3.3.0
# eyeD3==0.8.10
filelock==3.6.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
# grpcio==1.30.2
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.7.0
idna==3.4
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
# louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
# nemo-emblems==5.8.0
# netaddr==0.8.0
# netifaces==0.11.0
oauthlib==3.2.2
# onboard==1.4.1
package-name==0.1
packaging==21.3
# PAM==0.4.2
# pexpect==4.8.0
Pillow==10.0.0
protobuf==4.23.3
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
# pycurl==7.44.1
pydantic==1.10.12
pyelftools==0.27
# PyGObject==3.42.1
# PyICU==2.8.1
# pyinotify==0.9.6
# PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.0
pyparted==3.11.7
pyRFC3339==1.1
# python-apt==2.4.0+ubuntu2
# python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
uritemplate==4.1.1
urllib3==1.26.16
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
